+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Change laser rendering to be compatible with Iris
 * Fix machines in the Engineer House being obtainable when broken
 * Fixed the drill breaking when out of energy.
+* Hopefully fixed crash with Lithium on dedicated servers.
 
 # 1.3.8+mc1.21.1
 * Lock the Immersive Engineering cloche compat recipes behind a mod loaded condition